Alu/requirements.txt
Flask==2.0.3
Werkzeug==2.2.2
PyPDF2==3.0.1
Requests==2.32.3
pdoc
pipreqs
llama-cpp-python
huggingface-hub